src/ai_scheduler.py
```python
import calendar
from src.nlp_engine import NLP_Engine
from src.email_integration import EmailIntegration
from src.user_account import UserAccount

class AIScheduler:
    def __init__(self):
        self.nlp_engine = NLP_Engine()
        self.email_integration = EmailIntegration()
        self.user_account = UserAccount()
        self.calendar_data = {}
        self.user_preferences = {}

    def integrate_calendar(self, calendar_system):
        # Code to integrate with the specified calendar system
        pass

    def schedule_task(self, user_query):
        # Extract relevant information from user query using NLP engine
        task_details = self.nlp_engine.extract_task_details(user_query)

        # Check user's availability and preferences
        user_availability = self.check_availability(task_details)
        if user_availability:
            # Schedule the task
            self.calendar_data = self.update_calendar(task_details)
            return "scheduleSuccess"
        else:
            return "scheduleFailure"

    def check_availability(self, task_details):
        # Code to check user's availability based on task details and user preferences
        pass

    def update_calendar(self, task_details):
        # Code to update calendar data with the new task
        pass

    def manage_email(self, email_data):
        # Code to manage email interactions
        self.email_integration.manage_email(email_data)

    def update_preferences(self, user_preferences):
        # Code to update user preferences
        self.user_preferences = user_preferences
```
